refactoring: go fmt, reorg
