+++ weed/filer2/filechunks.go
@@ -82,10 +82,10 @@ func ReadFromChunks(chunks []*filer_pb.FileChunk, offset int64, size int) (views
 
 func logPrintf(name string, visibles []*visibleInterval) {
 	/*
-	log.Printf("%s len %d", name, len(visibles))
-	for _, v := range visibles {
-		log.Printf("%s:  => %+v", name, v)
-	}
+		log.Printf("%s len %d", name, len(visibles))
+		for _, v := range visibles {
+			log.Printf("%s:  => %+v", name, v)
+		}
 	*/
 }
